@@ -29,7 +29,7 @@ under the License.
 
     <groupId>org.apache.flink</groupId>
     <artifactId>flink-kubernetes-operator-parent</artifactId>
-    <version>1.11-SNAPSHOT</version>
+    <version>1.13-SNAPSHOT</version>
 
     <name>Flink Kubernetes: </name>
     <packaging>pom</packaging>
@@ -66,8 +66,8 @@ under the License.
     </modules>
 
     <properties>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
         <maven-jar-plugin.version>3.3.0</maven-jar-plugin.version>
         <maven-surefire-plugin.version>3.0.0-M4</maven-surefire-plugin.version>
         <maven-failsafe-plugin.version>3.0.0-M5</maven-failsafe-plugin.version>
@@ -75,10 +75,10 @@ under the License.
         <maven-javadoc-plugin.version>3.3.2</maven-javadoc-plugin.version>
         <git-commit-id-maven-plugin.version>5.0.0</git-commit-id-maven-plugin.version>
 
-        <operator.sdk.version>4.9.4</operator.sdk.version>
-        <operator.sdk.webhook-framework.version>1.1.1</operator.sdk.webhook-framework.version>
+        <operator.sdk.version>5.1.1</operator.sdk.version>
+        <operator.sdk.webhook-framework.version>3.0.0</operator.sdk.webhook-framework.version>
 
-        <fabric8.version>6.13.2</fabric8.version>
+        <fabric8.version>7.3.1</fabric8.version>
 
         <lombok.version>1.18.30</lombok.version>
         <commons-lang3.version>3.12.0</commons-lang3.version>
@@ -138,6 +138,10 @@ under the License.
                 <artifactId>kubernetes-client</artifactId>
                 <version>${fabric8.version}</version>
                 <exclusions>
+                    <exclusion>
+                        <groupId>io.fabric8</groupId>
+                        <artifactId>kubernetes-httpclient-vertx</artifactId>
+                    </exclusion>
                     <exclusion>
                         <groupId>com.squareup.okhttp3</groupId>
                         <artifactId>*</artifactId>
@@ -489,7 +493,7 @@ under the License.
             <plugin>
                 <groupId>org.cyclonedx</groupId>
                 <artifactId>cyclonedx-maven-plugin</artifactId>
-                <version>2.7.9</version>
+                <version>2.9.0</version>
                 <executions>
                     <execution>
                         <phase>package</phase>
