@@ -52,6 +52,7 @@ import org.apache.flink.kubernetes.operator.service.CheckpointHistoryWrapper;
 import org.apache.flink.kubernetes.operator.service.SuspendMode;
 import org.apache.flink.kubernetes.operator.standalone.StandaloneKubernetesConfigOptionsInternal;
 import org.apache.flink.runtime.client.JobStatusMessage;
+import org.apache.flink.runtime.clusterframework.ApplicationStatus;
 import org.apache.flink.runtime.execution.ExecutionState;
 import org.apache.flink.runtime.jobgraph.SavepointConfigOptions;
 import org.apache.flink.runtime.jobmanager.HighAvailabilityMode;
@@ -63,6 +64,7 @@ import org.apache.flink.runtime.messages.webmonitor.JobDetails;
 import org.apache.flink.runtime.messages.webmonitor.MultipleJobsDetails;
 import org.apache.flink.runtime.rest.messages.DashboardConfiguration;
 import org.apache.flink.runtime.rest.messages.EmptyResponseBody;
+import org.apache.flink.runtime.rest.messages.JobExceptionsInfoWithHistory;
 import org.apache.flink.runtime.rest.messages.JobsOverviewHeaders;
 import org.apache.flink.runtime.rest.messages.job.metrics.AggregatedMetric;
 import org.apache.flink.runtime.rest.messages.job.metrics.AggregatedMetricsResponseBody;
@@ -98,6 +100,7 @@ import java.util.Optional;
 import java.util.Random;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeoutException;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
@@ -111,6 +114,7 @@ public class TestingFlinkService extends AbstractFlinkService {
                     "15.0.0",
                     DashboardConfiguration.FIELD_NAME_FLINK_REVISION,
                     "1234567 @ 1970-01-01T00:00:00+00:00");
+    private final Map<JobID, JobExceptionsInfoWithHistory> jobExceptionsMap = new HashMap<>();
     public static final String SNAPSHOT_ERROR_MESSAGE = "Failed";
 
     private int savepointCounter = 0;
@@ -129,6 +133,7 @@ public class TestingFlinkService extends AbstractFlinkService {
     @Setter private boolean checkpointAvailable = true;
     @Setter private boolean jobManagerReady = true;
     @Setter private boolean deployFailure = false;
+    @Setter private Exception makeItFailWith;
     @Setter private boolean triggerSavepointFailure = false;
     @Setter private boolean disposeSavepointFailure = false;
     @Setter private Runnable sessionJobSubmittedCallback;
@@ -139,6 +144,7 @@ public class TestingFlinkService extends AbstractFlinkService {
     @Getter private final Map<String, Boolean> checkpointTriggers = new HashMap<>();
     private final Map<Long, String> checkpointStats = new HashMap<>();
     @Setter private boolean throwCheckpointingDisabledError = false;
+    @Setter private Throwable jobFailedErr;
 
     @Getter private int desiredReplicas = 0;
     @Getter private int cancelJobCallCount = 0;
@@ -211,6 +217,9 @@ public class TestingFlinkService extends AbstractFlinkService {
     }
 
     protected void deployApplicationCluster(JobSpec jobSpec, Configuration conf) throws Exception {
+        if (makeItFailWith != null) {
+            throw makeItFailWith;
+        }
         if (deployFailure) {
             throw new Exception("Deployment failure");
         }
@@ -269,6 +278,10 @@ public class TestingFlinkService extends AbstractFlinkService {
             @Nullable String savepoint)
             throws Exception {
 
+        if (makeItFailWith != null) {
+            throw makeItFailWith;
+        }
+
         if (deployFailure) {
             throw new Exception("Deployment failure");
         }
@@ -292,7 +305,33 @@ public class TestingFlinkService extends AbstractFlinkService {
         if (!isPortReady) {
             throw new TimeoutException("JM port is unavailable");
         }
+
+        if (jobFailedErr != null) {
+            return Optional.of(new JobStatusMessage(jobID, "n", JobStatus.FAILED, 0));
+        }
+
         return super.getJobStatus(conf, jobID);
+    }
+
+    @Override
+    public JobExceptionsInfoWithHistory getJobExceptions(
+            AbstractFlinkResource resource, JobID jobId, Configuration observeConfig) {
+        return jobExceptionsMap.getOrDefault(jobId, null);
+    }
+
+    @Override
+    public JobResult requestJobResult(Configuration conf, JobID jobID) throws Exception {
+        if (jobFailedErr != null) {
+            return new JobResult.Builder()
+                    .jobId(jobID)
+                    .serializedThrowable(new SerializedThrowable(jobFailedErr))
+                    .netRuntime(1)
+                    .accumulatorResults(new HashMap<>())
+                    .applicationStatus(ApplicationStatus.FAILED)
+                    .build();
+        }
+
+        return super.requestJobResult(conf, jobID);
     }
 
     public List<Tuple3<String, JobStatusMessage, Configuration>> listJobs() {
@@ -463,11 +502,13 @@ public class TestingFlinkService extends AbstractFlinkService {
     }
 
     private MultipleJobsDetails getMultipleJobsDetails() {
-        return new MultipleJobsDetails(
-                jobs.stream()
-                        .map(tuple -> tuple.f1)
-                        .map(TestingFlinkService::toJobDetails)
-                        .collect(Collectors.toList()));
+        MultipleJobsDetails multipleJobsDetails =
+                new MultipleJobsDetails(
+                        jobs.stream()
+                                .map(tuple -> tuple.f1)
+                                .map(TestingFlinkService::toJobDetails)
+                                .collect(Collectors.toList()));
+        return multipleJobsDetails;
     }
 
     private AggregatedMetricsResponseBody getSubtaskMetrics() {
@@ -598,8 +639,9 @@ public class TestingFlinkService extends AbstractFlinkService {
                     Optional<CheckpointHistoryWrapper.PendingCheckpointInfo>>
             getCheckpointInfo(JobID jobId, Configuration conf) throws Exception {
         if (throwCheckpointingDisabledError) {
-            throw new RestClientException(
-                    "Checkpointing has not been enabled", HttpResponseStatus.BAD_REQUEST);
+            throw new ExecutionException(
+                    new RestClientException(
+                            "Checkpointing has not been enabled", HttpResponseStatus.BAD_REQUEST));
         }
 
         if (checkpointInfo != null) {
@@ -692,5 +734,39 @@ public class TestingFlinkService extends AbstractFlinkService {
     public Map<String, String> getMetrics(
             Configuration conf, String jobId, List<String> metricNames) {
         return metricsValues;
+    }
+
+    /**
+     * Utilities to add exception history for testing.
+     *
+     * @param jobId of Job for which exception history is mocked
+     * @param exceptionName dummy exception name
+     * @param stackTrace dummy stack trace
+     * @param timestamp dummy timestamp
+     */
+    public void addExceptionHistory(
+            JobID jobId, String exceptionName, String stackTrace, long timestamp) {
+        JobExceptionsInfoWithHistory.RootExceptionInfo exception =
+                new JobExceptionsInfoWithHistory.RootExceptionInfo(
+                        exceptionName,
+                        stackTrace,
+                        timestamp,
+                        Map.of("label-key", "label-value"),
+                        List.of() // concurrentExceptions
+                        );
+
+        JobExceptionsInfoWithHistory existing = jobExceptionsMap.get(jobId);
+        List<JobExceptionsInfoWithHistory.RootExceptionInfo> entries =
+                existing != null && existing.getExceptionHistory() != null
+                        ? new ArrayList<>(existing.getExceptionHistory().getEntries())
+                        : new ArrayList<>();
+        entries.add(exception);
+
+        JobExceptionsInfoWithHistory.JobExceptionHistory exceptionHistory =
+                new JobExceptionsInfoWithHistory.JobExceptionHistory(entries, false);
+
+        JobExceptionsInfoWithHistory newExceptionHistory =
+                new JobExceptionsInfoWithHistory(exceptionHistory);
+        jobExceptionsMap.put(jobId, newExceptionHistory);
     }
 }
