@@ -30,15 +30,14 @@ import org.apache.flink.kubernetes.operator.exception.MissingJobManagerException
 import org.apache.flink.kubernetes.operator.observer.AbstractFlinkResourceObserver;
 import org.apache.flink.kubernetes.operator.reconciler.ReconciliationUtils;
 import org.apache.flink.kubernetes.operator.utils.EventRecorder;
+import org.apache.flink.kubernetes.operator.utils.EventUtils;
 import org.apache.flink.kubernetes.operator.utils.FlinkUtils;
 
-import io.fabric8.kubernetes.api.model.ContainerStateWaiting;
 import io.fabric8.kubernetes.api.model.ContainerStatus;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodList;
 import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.fabric8.kubernetes.api.model.apps.DeploymentCondition;
-import io.fabric8.kubernetes.api.model.apps.DeploymentSpec;
 import io.fabric8.kubernetes.api.model.apps.DeploymentStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -46,7 +45,7 @@ import org.slf4j.LoggerFactory;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
+import java.util.stream.Stream;
 
 /** Base observer for session and application clusters. */
 public abstract class AbstractFlinkDeploymentObserver
@@ -117,11 +116,11 @@ public abstract class AbstractFlinkDeploymentObserver
                 ctx.getJosdkContext().getSecondaryResource(Deployment.class);
         if (deployment.isPresent()) {
             DeploymentStatus status = deployment.get().getStatus();
-            DeploymentSpec spec = deployment.get().getSpec();
             if (status != null
                     && status.getAvailableReplicas() != null
-                    && spec.getReplicas().intValue() == status.getReplicas()
-                    && spec.getReplicas().intValue() == status.getAvailableReplicas()
+                    // One available JM is enough to run the job correctly
+                    && status.getReplicas() > 0
+                    && status.getAvailableReplicas() > 0
                     && ctx.getFlinkService().isJobManagerPortReady(ctx.getObserveConfig())) {
 
                 // typically it takes a few seconds for the REST server to be ready
@@ -135,7 +134,7 @@ public abstract class AbstractFlinkDeploymentObserver
             try {
                 checkFailedCreate(status);
                 // checking the pod is expensive; only do it when the deployment isn't ready
-                checkContainerBackoff(ctx);
+                checkContainerErrors(ctx);
             } catch (DeploymentFailedException dfe) {
                 // throw only when not already in error status to allow for spec update
                 deploymentStatus.getJobStatus().setState(JobStatus.RECONCILING);
@@ -172,21 +171,28 @@ public abstract class AbstractFlinkDeploymentObserver
         }
     }
 
-    private void checkContainerBackoff(FlinkResourceContext<FlinkDeployment> ctx) {
+    private void checkContainerErrors(FlinkResourceContext<FlinkDeployment> ctx) {
         PodList jmPods =
                 ctx.getFlinkService().getJmPodList(ctx.getResource(), ctx.getObserveConfig());
         for (Pod pod : jmPods.getItems()) {
-            for (ContainerStatus cs : pod.getStatus().getContainerStatuses()) {
-                ContainerStateWaiting csw = cs.getState().getWaiting();
-                if (csw != null
-                        && Set.of(
-                                        DeploymentFailedException.REASON_CRASH_LOOP_BACKOFF,
-                                        DeploymentFailedException.REASON_IMAGE_PULL_BACKOFF,
-                                        DeploymentFailedException.REASON_ERR_IMAGE_PULL)
-                                .contains(csw.getReason())) {
-                    throw new DeploymentFailedException(csw);
-                }
-            }
+            var podStatus = pod.getStatus();
+            Stream.concat(
+                            podStatus.getContainerStatuses().stream(),
+                            podStatus.getInitContainerStatuses().stream())
+                    .forEach(AbstractFlinkDeploymentObserver::checkContainerError);
+
+            // No obvious errors were found, check for volume mount issues
+            EventUtils.checkForVolumeMountErrors(ctx.getKubernetesClient(), pod);
+        }
+    }
+
+    private static void checkContainerError(ContainerStatus cs) {
+        if (cs.getState() == null || cs.getState().getWaiting() == null) {
+            return;
+        }
+        if (DeploymentFailedException.CONTAINER_ERROR_REASONS.contains(
+                cs.getState().getWaiting().getReason())) {
+            throw DeploymentFailedException.forContainerStatus(cs);
         }
     }
 
