@@ -35,6 +35,7 @@ import org.apache.flink.kubernetes.operator.controller.FlinkResourceContext;
 import org.apache.flink.kubernetes.operator.exception.UpgradeFailureException;
 import org.apache.flink.kubernetes.operator.health.ClusterHealthInfo;
 import org.apache.flink.kubernetes.operator.observer.ClusterHealthEvaluator;
+import org.apache.flink.kubernetes.operator.observer.ClusterHealthResult;
 import org.apache.flink.kubernetes.operator.reconciler.ReconciliationUtils;
 import org.apache.flink.kubernetes.operator.service.FlinkService;
 import org.apache.flink.kubernetes.operator.service.SuspendMode;
@@ -223,10 +224,12 @@ public class ApplicationReconciler
     @Override
     protected boolean cancelJob(FlinkResourceContext<FlinkDeployment> ctx, SuspendMode suspendMode)
             throws Exception {
+        var cancelTs = Instant.now();
         var result =
                 ctx.getFlinkService()
                         .cancelJob(ctx.getResource(), suspendMode, ctx.getObserveConfig());
-        result.getSavepointPath().ifPresent(location -> setUpgradeSavepointPath(ctx, location));
+        result.getSavepointPath()
+                .ifPresent(location -> setUpgradeSavepointPath(ctx, location, cancelTs));
         return result.isPending();
     }
 
@@ -266,6 +269,9 @@ public class ApplicationReconciler
 
         var deployment = ctx.getResource();
         var observeConfig = ctx.getObserveConfig();
+        var clusterHealthInfo =
+                ClusterHealthEvaluator.getLastValidClusterHealthInfo(
+                        deployment.getStatus().getClusterInfo());
         boolean shouldRestartJobBecauseUnhealthy =
                 shouldRestartJobBecauseUnhealthy(deployment, observeConfig);
         boolean shouldRecoverDeployment = shouldRecoverDeployment(observeConfig, deployment);
@@ -286,7 +292,10 @@ public class ApplicationReconciler
                         EventRecorder.Type.Warning,
                         EventRecorder.Reason.RestartUnhealthyJob,
                         EventRecorder.Component.Job,
-                        MSG_RESTART_UNHEALTHY,
+                        Optional.ofNullable(clusterHealthInfo)
+                                .map(ClusterHealthInfo::getHealthResult)
+                                .map(ClusterHealthResult::getError)
+                                .orElse(MSG_RESTART_UNHEALTHY),
                         ctx.getKubernetesClient());
                 cleanupAfterFailedJob(ctx);
             }
@@ -310,7 +319,7 @@ public class ApplicationReconciler
                     ClusterHealthEvaluator.getLastValidClusterHealthInfo(clusterInfo);
             if (clusterHealthInfo != null) {
                 LOG.debug("Cluster info contains job health info");
-                if (!clusterHealthInfo.isHealthy()) {
+                if (!clusterHealthInfo.getHealthResult().isHealthy()) {
                     if (deployment.getSpec().getJob().getUpgradeMode() == UpgradeMode.STATELESS) {
                         LOG.debug("Stateless job, recovering unhealthy jobmanager deployment");
                         restartNeeded = true;
