@@ -65,6 +65,7 @@ import org.apache.flink.kubernetes.operator.config.KubernetesOperatorConfigOptio
 import org.apache.flink.kubernetes.operator.exception.UpgradeFailureException;
 import org.apache.flink.kubernetes.operator.health.ClusterHealthInfo;
 import org.apache.flink.kubernetes.operator.observer.ClusterHealthEvaluator;
+import org.apache.flink.kubernetes.operator.observer.ClusterHealthResult;
 import org.apache.flink.kubernetes.operator.reconciler.ReconciliationUtils;
 import org.apache.flink.kubernetes.operator.reconciler.SnapshotType;
 import org.apache.flink.kubernetes.operator.reconciler.TestReconcilerAdapter;
@@ -98,6 +99,7 @@ import java.time.Clock;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.ZoneId;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -127,7 +129,6 @@ import static org.apache.flink.kubernetes.operator.reconciler.SnapshotType.CHECK
 import static org.apache.flink.kubernetes.operator.reconciler.SnapshotType.SAVEPOINT;
 import static org.apache.flink.kubernetes.operator.reconciler.deployment.AbstractFlinkResourceReconciler.MSG_SUBMIT;
 import static org.apache.flink.kubernetes.operator.reconciler.deployment.ApplicationReconciler.MSG_RECOVERY;
-import static org.apache.flink.kubernetes.operator.reconciler.deployment.ApplicationReconciler.MSG_RESTART_UNHEALTHY;
 import static org.apache.flink.kubernetes.operator.utils.SnapshotUtils.getLastSnapshotStatus;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -302,6 +303,13 @@ public class ApplicationReconcilerTest extends OperatorTestBase {
         reconciler.reconcile(statefulUpgrade, context);
 
         assertEquals(0, flinkService.getRunningCount());
+
+        var spInfo = statefulUpgrade.getStatus().getJobStatus().getSavepointInfo();
+        assertEquals("savepoint_0", spInfo.getLastSavepoint().getLocation());
+        assertEquals(SnapshotTriggerType.UPGRADE, spInfo.getLastSavepoint().getTriggerType());
+        assertEquals(
+                spInfo.getLastSavepoint(),
+                new LinkedList<>(spInfo.getSavepointHistory()).getLast());
 
         reconciler.reconcile(statefulUpgrade, context);
 
@@ -1215,12 +1223,11 @@ public class ApplicationReconcilerTest extends OperatorTestBase {
         var clusterHealthInfo = new ClusterHealthInfo();
         clusterHealthInfo.setTimeStamp(System.currentTimeMillis());
         clusterHealthInfo.setNumRestarts(2);
-        clusterHealthInfo.setHealthy(false);
+        clusterHealthInfo.setHealthResult(ClusterHealthResult.error("error"));
         ClusterHealthEvaluator.setLastValidClusterHealthInfo(
                 deployment.getStatus().getClusterInfo(), clusterHealthInfo);
         reconciler.reconcile(deployment, context);
-        Assertions.assertEquals(
-                MSG_RESTART_UNHEALTHY, flinkResourceEventCollector.events.remove().getMessage());
+        Assertions.assertEquals("error", flinkResourceEventCollector.events.remove().getMessage());
     }
 
     @Test
